Fix #48 moved all trip friends api call to user module
